In the step2 branch, more libraries were introduced, such as: Navigation - to implement more convenient navigation through objects Picasso - for displaying received images from requests
In the Domain layer, two classes were implemented, ApiService for sending requests, a single tone object ApiFactory which contains the retrofit implementation and the base URL of the site. The Mapper class was also added, which converts objects received from the site into the desired Post class, and also converts time into a string.

Two new fragments and the corresponding ViewModel were added to the Presentation layer. The first fragment is responsible for displaying the list of posts and interacting with it. The second one is needed to display the image in a larger view. A package containing an adapter implementation for RecyclerView was also added.